Minor API change and docs update
- `TargetSize` moved from utils.py to mydia.py, which makes much more sense as cannot be categorized as a utility.
- Docs update
- Updated README.rst to reflect the updates in the API
