+++ mydia/__init__.py
@@ -1,9 +1,2 @@
 from .mydia import *
-from .utils import (
-    TargetSize,
-    _mode_auto,
-    _mode_first,
-    _mode_last,
-    _mode_middle,
-    _mode_random,
-)
+from .utils import _mode_auto, _mode_first, _mode_last, _mode_middle, _mode_random
